+++ Runtime/GsplatAssetSpark.cs
@@ -9,7 +9,6 @@ using Unity.Mathematics;
 using UnityEngine;
 using UnityEngine.Rendering;
 
-
 namespace Gsplat
 {
     /// <summary>
@@ -110,6 +109,9 @@ namespace Gsplat
         public override void SetupMaterialPropertyBlock(MaterialPropertyBlock propertyBlock,
             GsplatResource resource)
         {
+            var cs = GsplatMaterial.InitOrderShader;
+            m_kernelInitOrder = cs.FindKernel("InitOrder");
+
             var res = (GsplatResourceSpark)resource;
             propertyBlock.SetBuffer(k_packedSplatsBuffer, res.PackedSplatsBuffer);
             if (SHBands >= 1)
@@ -120,11 +122,11 @@ namespace Gsplat
                 propertyBlock.SetBuffer(k_packedSH3Buffer, res.PackedSH3Buffer);
         }
 
-        public override void ComputeDepth(GsplatMaterial material, CommandBuffer cmd, Matrix4x4 matrixMv,
+        public override void ComputeDepth(CommandBuffer cmd, Matrix4x4 matrixMv,
             ISorterResource sorterResource, GsplatResource resource)
         {
             var res = (GsplatResourceSpark)resource;
-            var cs = material.CalcDepthShader;
+            var cs = GsplatMaterial.CalcDepthShader;
             const int kernelCalcDepthSpark = 0;
             cmd.SetComputeIntParam(cs, k_splatCount, (int)SplatCount);
             cmd.SetComputeMatrixParam(cs, k_matrixMv, matrixMv);
@@ -132,6 +134,25 @@ namespace Gsplat
             cmd.SetComputeBufferParam(cs, kernelCalcDepthSpark, k_depthBuffer, sorterResource.InputKeys);
             cmd.SetComputeBufferParam(cs, kernelCalcDepthSpark, k_orderBuffer, sorterResource.OrderBuffer);
             cmd.DispatchCompute(cs, kernelCalcDepthSpark, (int)GsplatUtils.DivRoundUp(SplatCount, 1024), 1, 1);
+        }
+
+        public override void InitOrder(ISorterResource sorterResource, GsplatResource resource, bool updateBounds)
+        {
+            var cs = GsplatMaterial.InitOrderShader;
+            var res = (GsplatResourceSpark)resource;
+
+            sorterResource.OrderBuffer.SetCounterValue(0);
+
+            uint threadBlocks = GsplatUtils.DivRoundUp(SplatCount, 1024);
+
+            cs.SetInt(k_splatCount, (int)SplatCount);
+            cs.SetBuffer(m_kernelInitOrder, k_orderBuffer, sorterResource.OrderBuffer);
+            cs.SetBuffer(m_kernelInitOrder, k_packedSplatsBuffer, res.PackedSplatsBuffer);
+            if (updateBounds)
+                cs.EnableKeyword("UPDATE_BOUNDS");
+            else
+                cs.DisableKeyword("UPDATE_BOUNDS");
+            cs.Dispatch(m_kernelInitOrder, (int)threadBlocks, 1, 1);
         }
 
         public override void LoadFromPly(string plyPath, ProgressCallback progressCallback = null)