+++ Runtime/GsplatResource.cs
@@ -57,7 +57,7 @@ namespace Gsplat
         public GraphicsBuffer PackedSH2Buffer { get; private set; }
         public GraphicsBuffer PackedSH3Buffer { get; private set; }
 
-        public GsplatResourceSpark(uint splatCount, byte shBands)
+        public GsplatResourceSpark(uint splatCount, byte shBands) : base()
         {
             if (splatCount == 0)
                 return;